removed image controller
