+++ src/main/resources/templates/imageuploadform.html
@@ -1,36 +1,42 @@
 <!doctype html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
-    <meta charset="UTF-8">
-    <title>SoundMachine Musician File Upload</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
+<meta charset="UTF-8">
+<title>SoundMachine Musician File Upload</title>
+<link rel="stylesheet"
+	href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
 
-    <link href="https://fonts.googleapis.com/css?family=Allerta+Stencil%7CJosefin+Sans%7CReem+Kufi&display=swap"
-        rel="stylesheet">
+<link
+	href="https://fonts.googleapis.com/css?family=Allerta+Stencil%7CJosefin+Sans%7CReem+Kufi&display=swap"
+	rel="stylesheet">
 
-    <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
-        integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n"
-        crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
-        integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
-        crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
-        integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6"
-        crossorigin="anonymous"></script>
+<script src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
+	integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n"
+	crossorigin="anonymous"></script>
+<script
+	src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
+	integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
+	crossorigin="anonymous"></script>
+<script
+	src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
+	integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6"
+	crossorigin="anonymous"></script>
 
-    <!-- zum lokalen Testen -->
-    <!--  <link rel="stylesheet" href="../static/css/style.css"> -->
-    <link rel="stylesheet" th:href="@{/css/style.css}">
+<!-- zum lokalen Testen -->
+<!--  <link rel="stylesheet" href="../static/css/style.css"> -->
+<link rel="stylesheet" th:href="@{/css/style.css}">
 
 </head>
 <body>
-	<!-- If musician.id is set, write edit, otherwise create -->
+	<header th:replace="index :: navigation_bar">Navigation</header>
 	<h3>Band Photo Upload</h3>
-	<form method="post" th:action="@{'/band/' + ${band.getId()} + '/changeimage'}" enctype = "multipart/form-data">
-		<label class="control-label">Wähle Bilddatei</label>
-		<input id="imagefile" name="imagefile" type="file" class="file" />
-		<br>
+	<form method="post"
+		th:action="@{'/band/' + ${band.getId()} + '/uploadimage'}"
+		enctype="multipart/form-data">
+		<label class="control-label">Wähle Bilddatei</label> <input
+			id="imagefile" name="imagefile" type="file" class="file" /> <br>
 		<button type="submit" class="btn btn-primary">Upload</button>
 	</form>
+	<footer th:replace="index :: footer">Footer</footer>
 </body>
 </html>
